tries to make preview only run on request (not in ci)
